test: Switch to vanilla CentOS 7
This switches off of the F1 instance type to a T3.XLarge which is
approximately 1/10th of the price.  At this stage of the game we're
working on optimizing the basic setup of the machine and the compilation
of the GNU tools from source.

The configuration of the machine is ocurring in the file
`cloud-config.yaml`.  For more information on the cloud-config format
consult the documentation:

https://cloudinit.readthedocs.io/en/20.4.1/topics/examples.html